% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_error.R
\name{check_credentials}
\alias{check_credentials}
\title{Check credentials from sweetpotatobase users.}
\usage{
check_credentials(dbname = "sweetpotatobase", user = "obenites",
  password = "dasdfsdgs",
  urltoken = "sgn:eggplant@sweetpotatobase-test.sgn.cornell.edu/brapi/v1/token")
}
\arguments{
\item{dbname}{character Database name. Currently, it only works with SOL genomics databases.}

\item{user}{character User name}

\item{password}{character Password}

\item{urltoken}{character \code{BRAPI} call URL  to  login in Sol Genomic databases.}
}
\description{
FieldbookApp data is captured by mobiles phones or tablets. After exporting this information, it should be read in R
in order to process, check and curate.
}
\author{
Omar Benites
}
