% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-fb.R
\name{check_fbapp}
\alias{check_fbapp}
\title{Check error in FieldbookApp data}
\usage{
check_fbapp(dfr)
}
\arguments{
\item{dfr}{data frame. FieldbookApp data.}
}
\description{
FieldbookApp data is captured by mobiles phones or tablets. After exporting this information, it should be read in R
in order to process, check and curate.
}
\author{
Omar Benites
}
