% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-fb.R
\name{ck_duplicate}
\alias{ck_duplicate}
\title{Check duplicates entries in FieldBookApp data}
\usage{
ck_duplicate(dfr = data.frame(), attribute)
}
\arguments{
\item{dfr}{data frame. FieldBookApp data.}

\item{attribute}{character. Column name for checking whether duplicates values exist in FieldBookApp data}
}
\description{
Check duplicated entries in specific attributes, ex. \code{"plot_name"}, in FieldBookApp data
}
\examples{
dfr<- data.frame(plot_name= paste0("plot_","rep_",1:3), value= 1:3)
ck_duplicate(dfr,"plot_name")
}
\author{
Omar Benites
}
