% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fbapp2hidap}
\alias{fbapp2hidap}
\title{FieldBookApp Data Processing --------------------------------------------}
\usage{
fbapp2hidap(fieldbook)
}
\arguments{
\item{fieldbook}{fieldbook from FieldBookApp to HIDAP}
}
\description{
Return pvs form parameters
}
\author{
Omar Benites
}
