% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fbapp2json}
\alias{fbapp2json}
\title{Convert FieldbookApp data to json structures}
\usage{
fbapp2json(dfr, token = "lfsermmo93;3r")
}
\arguments{
\item{dfr}{data.frame}

\item{token}{character Token is provided by SOL genomics databases through \code{BRAPI} calls.}
}
\description{
FieldbookApp files (csv) should be transformed into json files in order to upload into Sol genomics databases.
}
\author{
Omar Benites
}
