% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbcheck_server.R
\name{fbcheck_server}
\alias{fbcheck_server}
\title{Server component for traittools in fbcheck}
\usage{
fbcheck_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{nameo of the output element}

\item{session}{shinyserver session}

\item{values}{The reactive values}
}
\description{
Returns server side components
}
\author{
Omar Benites
}

