% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbcheck_server_sbase.R
\name{fbcheck_server_sbase}
\alias{fbcheck_server_sbase}
\title{Server component for traittools and data quality for HIDAP-SweetPotatoBase}
\usage{
fbcheck_server_sbase(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{nameo of the output element}

\item{session}{shinyserver session}

\item{values}{The reactive values}
}
\description{
Returns server side components for HIDAP-SweetPotatoBase
}
\author{
Omar Benites
}
