% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbcheck_ui.R
\name{fbcheck_ui}
\alias{fbcheck_ui}
\title{UI for traittools
Returns user friendly ui}
\usage{
fbcheck_ui(type = "tab", title = "Data Quality and Processing",
  name = "data_processing")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for traittools
Returns user friendly ui
}
\author{
Omar Benites
}

