% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbcheck_ui_sbase.R
\name{fbcheck_ui_sbase}
\alias{fbcheck_ui_sbase}
\title{UI for traittools and data quality for HIDAP-SweetPotatoBase
Returns user friendly ui for HIDAP-SweetPotatoBase}
\usage{
fbcheck_ui_sbase(type = "tab", title = "Data Quality and Processing",
  name = "data_processing")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{display title name}

\item{name}{UI TabName}
}
\description{
UI for traittools and data quality for HIDAP-SweetPotatoBase
Returns user friendly ui for HIDAP-SweetPotatoBase
}
\author{
Omar Benites
}
