% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{form_parameters}
\alias{form_parameters}
\title{Get form parameters from PVS forms}
\usage{
form_parameters(list_form)
}
\arguments{
\item{list_form}{list of forms}
}
\description{
Return pvs form parameters
}
\author{
Omar Benites
}
