% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_function.R
\name{get.fb.param}
\alias{get.fb.param}
\title{Function to obtain parameter from fieldbook}
\usage{
get.fb.param(fp, sheet, param)
}
\arguments{
\item{fp}{fieldbook path}

\item{sheet}{fieldbook's sheet}

\item{param}{Parameters}
}
\description{
This function gets parameters or values from fieldbook excel file. Do an excel scrapping.
}

