% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hidap2fbApp}
\alias{hidap2fbApp}
\title{FieldBookApp Data Processing --------------------------------------------}
\usage{
hidap2fbApp(fieldbook)
}
\arguments{
\item{fieldbook}{fieldbook from HIDAP to FieldBookApp}
}
\description{
Return pvs form parameters
}
\author{
Omar Benites
}
