% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{out_form_table}
\alias{out_form_table}
\title{output of the organoletic form in a table}
\usage{
out_form_table(form)
}
\arguments{
\item{form}{organoleptic form}
}
\description{
return an organized table
}
\author{
Omar Benites
}
