% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_function.R
\name{render_categorical}
\alias{render_categorical}
\title{Function for scale condition}
\usage{
render_categorical(scale_condition)
}
\arguments{
\item{scale_condition}{The scale of the trait.}
}
\description{
In plant traits, we generally find quantitative and cualitative scales. Then, the scale condition
creates javascript renders to higlight values which are out of range.
}
\author{
omar benites
}

