% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trait_function.R
\name{render_quantitative}
\alias{render_quantitative}
\title{Function for render quantitative traits in javascript conditions}
\usage{
render_quantitative(ll, ul)
}
\arguments{
\item{ll}{lower limit}

\item{ul}{upper limmit}
}
\description{
Function to get the scale of differents trait,
}
\author{
omar benites
}

