% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trait_function.R
\name{render_trait}
\alias{render_trait}
\title{R function for Javascript Render Trait}
\usage{
render_trait(trait, datadict)
}
\arguments{
\item{trait}{trait}

\item{datadict}{data dictionary}
}
\description{
Function to get the scale of differents trait,
}
\author{
omar benites
}

