% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rhandsontable_update}
\alias{rhandsontable_update}
\title{Update rhandsontable}
\usage{
rhandsontable_update(fieldbook)
}
\arguments{
\item{fieldbook}{data.frame field data trough rhandsontable}
}
\description{
Get updates from rhandonstable after user modifications
}
\author{
Omar Benites
}
