% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pvs.R
\docType{data}
\name{selcriteria}
\alias{selcriteria}
\title{Selection Criteria Data for HiDAP}
\format{A List Alist of selection criteria collected in participatory varietal selection}
\source{
International Potato Center, potato experimental data.
}
\usage{
selcriteria
}
\description{
This dataset contains Selection Criteria Data collected on Farm
}
\references{
This data is related to HiDAP crop template
}

