% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_tidy_form}
\alias{split_tidy_form}
\title{Split organoleptic forms}
\usage{
split_tidy_form(form)
}
\arguments{
\item{form}{organoleptic form}
}
\description{
split in tiny data frames all the orgaleptic forms
}
\author{
Omar Benites
}

