% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upload_studies}
\alias{upload_studies}
\title{Upload \code{FieldBookApp} data or studies from HIDAP to SweetPotatoBase trough Breeding API (BrAPI)}
\usage{
upload_studies(dbname = "sweetpotatobase",
  urltoken = "sgn:eggplant@sweetpotatobase-test.sgn.cornell.edu/brapi/v1/token",
  urlput = "sgn:eggplant@sweetpotatobase-test.sgn.cornell.edu/brapi/v1/observations",
  user = "obenites", password = ";c8U:G&z:X", dfr)
}
\arguments{
\item{dbname}{character Database name. Currently, it only works with SOL genomics databases.}

\item{urltoken}{character \code{BRAPI} call URL  to  login in Sol Genomic databases.}

\item{urlput}{character \code{BRAPI} call to \code{PUT} URL to studies.}

\item{user}{character User name}

\item{password}{character Password}

\item{dfr}{data.frame Fieldbook data collected by FieldBookApp.}
}
\description{
Plant breeders' tasks involve the collection, curation and preservation of field data. 
The FieldBookApp mobile application provide mobile-based data collection. Then, researchers
curate all this information in \code{HIDAP} (online and offline); and subsequently, 
they will upload each dataset to \code{SweetPotatoBase} (a \code{SolGenomics} Database). 
The latter mentioned is performed through the Breeding API (\code{BrAPI}) (see \code{https://brapi.docs.apiary.io}).
}
\author{
Omar Benites
}
