% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{x_form}
\alias{x_form}
\title{Get x values from PVS forms}
\usage{
x_form(form, genotypes = NA, name_panel = NA, n_panel = NA,
  sex_panel = NA)
}
\arguments{
\item{form}{pvs form}

\item{genotypes}{genotypes}

\item{name_panel}{the panelist name}

\item{n_panel}{the number of panelist}

\item{sex_panel}{the sex of the panelist}
}
\description{
Return the form with x marks
}
\author{
Omar Benites
}

