% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{x_values}
\alias{x_values}
\title{Return 'x' values (marks made by farmers in organoleptic forms)}
\usage{
x_values(vec, values)
}
\arguments{
\item{vec}{vector}

\item{values}{categorical values or scales}
}
\description{
get the x values from organoleptic forms
}
\author{
Omar Benites
}
