% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fbapp2hidap}
\alias{fbapp2hidap}
\title{Get form parameters from PVS forms}
\usage{
fbapp2hidap(fieldbook)
}
\arguments{
\item{fieldbook}{fieldbook from FieldBookApp}
}
\description{
Return pvs form parameters
}
\author{
Omar Benites
}
