% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trait_function.R
\name{render_categorical}
\alias{render_categorical}
\title{Function for scale condition}
\usage{
render_categorical(scale_condition)
}
\arguments{
\item{scale_condition}{}
}
\description{
Function to get the scale of differents trait,
}
\author{
omar benites
}

