% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trait_function.R
\name{trait_type}
\alias{trait_type}
\title{Function for trait type}
\usage{
trait_type(trait, datadict)
}
\arguments{
\item{trait}{trait}

\item{datadict}{data dictionary}
}
\description{
Function for trait type
}
\author{
omar benites
}

